refactor: modularize runner shared helpers
