+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared/__init__.py
@@ -1,0 +1,29 @@
+"""Shared helpers for runner modules."""
+from __future__ import annotations
+
+from .costs import estimate_cost, provider_model
+from .logging import (
+    MetricsPath,
+    error_family,
+    log_provider_call,
+    log_provider_skipped,
+    log_run_metric,
+    resolve_event_logger,
+)
+from .rate_limiter import RateLimiter, asyncio, resolve_rate_limiter, threading, time
+
+__all__ = [
+    "MetricsPath",
+    "resolve_event_logger",
+    "error_family",
+    "estimate_cost",
+    "provider_model",
+    "log_provider_skipped",
+    "log_provider_call",
+    "log_run_metric",
+    "RateLimiter",
+    "resolve_rate_limiter",
+    "asyncio",
+    "threading",
+    "time",
+]
