+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared/rate_limiter.py
@@ -1,0 +1,18 @@
+"""Rate limiter utilities exposed to runner modules."""
+from __future__ import annotations
+
+from .. import rate_limiter as _rate_limiter
+
+RateLimiter = _rate_limiter.RateLimiter
+resolve_rate_limiter = _rate_limiter.resolve_rate_limiter
+asyncio = _rate_limiter.asyncio
+threading = _rate_limiter.threading
+time = _rate_limiter.time
+
+__all__ = [
+    "RateLimiter",
+    "resolve_rate_limiter",
+    "asyncio",
+    "threading",
+    "time",
+]
